requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.24.0
opencv-python>=4.7.0
pyyaml>=6.0
scikit-learn>=1.2.0
timm>=0.9.2
facenet-pytorch>=0.2.1
matplotlib>=3.5.0